backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
openai==1.54.0
httpx==0.27.2
boto3==1.35.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.6.0
mangum==0.18.0
python-jose[cryptography]==3.3.0
pyjwt[crypto]==2.8.0
requests==2.31.0